+++ changelog.md
@@ -1,4 +1,14 @@
-Version 1.1
+Version 1.5
+- removed daily message showing casino income
+- added info panel section showing casino profit for both today and lifetime
+
+Version 1.4
+- reduced variance in amounts gambled
+- removed report for sol 0
+- removed report when there are no casinos built
+
+Version 1.3
+- Daily income of casinos printed to log file
 - Added option to show daily income of casinos
 - Added daily message for above option
 - Colonists bet more
@@ -6,4 +16,10 @@ Version 1.1
 - Gamblers play lower odds for higher payout
 - Tourists bet more
 - Watch out for geniuses
-- Option for who can gamble
+
+Version 1.2
+- added option to select who can gamble
+  - Everyone, Tourists, Humans, Martians, any combo
+ 
+Version 1.1
+- added gambling!